% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BFImage.R
\docType{class}
\name{BFImage-class}
\alias{BFImage}
\alias{BFImage-class}
\alias{as.Image.BFImage}
\alias{print.BFImage}
\alias{show,BFImage-method}
\title{Bio-Formats Image Class}
\usage{
BFImage(image = Image(), metadata = ImageMetadata(), omexml = "")

\S4method{show}{BFImage}(object)

\method{print}{BFImage}(x, short = FALSE, ...)

\method{as.Image}{BFImage}(x)
}
\arguments{
\item{image}{an Image object}

\item{metadata}{an ImageMetadata object containing image metadata}

\item{omexml}{a string containing a dumped OME-XML DOM tree}

\item{object}{A BFImage object}

\item{x}{a BFImage object}

\item{short}{turns off image data preview}

\item{...}{further arguments to be passed to other methods}
}
\description{
Extends the \code{\link[EBImage]{Image}} class from the \pkg{EBImage} package.
}
\section{Slots}{

\describe{
\item{\code{omexml}}{a string containing a dumped OME-XML DOM tree}
}}

