% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BioFormats.version}
\alias{BioFormats.version}
\title{Bio-Formats Version Number}
\usage{
BioFormats.version()
}
\description{
Provides the version of the Bio-Formats library embedded in the package.
}
\author{
Andrzej Oleś <\email{andrzej.oles@embl.de}>, 2014-2017
}
