% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BFImage.R
\name{dimorder}
\alias{dimorder}
\title{Get the ordering of the image frames}
\usage{
dimorder(x)
}
\arguments{
\item{x}{An Image object or an array}
}
\value{
A character vector giving the dimension labels.
}
\description{
Get the ordering of the image frames
}
\examples{
# sample timelapse image
img = read.image(mockFile(sizeC = 3, sizeT = 10))

dimorder(img)
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

