% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{FormatTools}
\alias{FormatTools}
\title{Bio-Formats FormatTools Class}
\format{An object of class \code{jclassName} of length 1.}
\usage{
FormatTools
}
\description{
A utility class for format reader and writer implementations.
}
\examples{
# List available pixel types
sapply(0:7, FormatTools$getPixelTypeString)
}
\author{
Andrzej Oleś <\email{andrzej.oles@embl.de}>, 2014-2017
}
\keyword{datasets}
