% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageMetadata.R
\docType{class}
\name{ImageMetadata-class}
\alias{ImageMetadata-class}
\alias{.ImageMetadata}
\alias{ImageMetadataList-class}
\alias{.ImageMetadataList}
\alias{ImageMetadata}
\alias{ImageMetadataList}
\alias{print.ImageMetadata}
\alias{show,ImageMetadata-method}
\alias{show,ImageMetadataList-method}
\alias{print.ImageMetadataList}
\title{ImageMetadata and ImageMetadataList Class}
\usage{
ImageMetadata(
  coreMetadata = NULL,
  globalMetadata = NULL,
  seriesMetadata = NULL
)

ImageMetadataList(...)

\method{print}{ImageMetadata}(x, list.len = 5L, ...)

\S4method{show}{ImageMetadata}(object)

\S4method{show}{ImageMetadataList}(object)

\method{print}{ImageMetadataList}(x, ...)
}
\arguments{
\item{coreMetadata}{a list of core metadata entries}

\item{globalMetadata}{a list of global metadata entries}

\item{seriesMetadata}{a list of series metadata entries}

\item{...}{further arguments to be passed to other methods}

\item{x}{An ImageMetadata object}

\item{list.len}{numeric; maximum number of metadata entries to display}

\item{object}{an \code{ImageMetadata} object}
}
\value{
\code{ImageMetadata} returns a new \linkS4class{ImageMetadata} object.

\code{ImageMetadataList} returns a new \linkS4class{ImageMetadataList} object.
}
\description{
Formal representation of image metadata.
}
\examples{
f = mockFile(sizeX = 256, sizeY = 256)
img = read.image(f)
metadata(img)
}
\author{
Andrzej Oleś <\email{andrzej.oles@gmail.com}>, 2014-2022
}
