% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BFImage.R
\name{coreMetadata}
\alias{coreMetadata}
\alias{globalMetadata}
\alias{seriesMetadata}
\title{Metadata Accessor Functions}
\usage{
coreMetadata(y, series = 1L, ...)

globalMetadata(y, series = 1L, ...)

seriesMetadata(y, series = 1L, ...)
}
\arguments{
\item{y}{A BFImage object or metadata list}

\item{series}{series ID}

\item{...}{arguments passed to \code{grep}}
}
\value{
Named list consisting of key value pairs.
}
\description{
Metadata Accessor Functions
}
\examples{
img = read.image(system.file("images", "nuclei.tif", package="EBImage"))

coreMetadata(img)

# subset for specific names

globalMetadata(img, pattern="Image")
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

