% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBioFormats.R
\docType{package}
\name{RBioFormats}
\alias{RBioFormats}
\title{R Interface to Bio-Formats}
\description{
Interfaces the \href{http://www.openmicroscopy.org/bio-formats}{Bio-Formats} Java library.
}
\references{
\href{http://www.openmicroscopy.org/bio-formats}{Bio-Formats website}
}
\author{
Andrzej Oleś <\email{andrzej.oles@gmail.com}>, 2014-2022
}
