% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataAccessors.R
\name{seriesCount}
\alias{seriesCount}
\title{Number of Image Series}
\usage{
seriesCount(x)
}
\arguments{
\item{x}{an \code{AnnotatedImageList} or \code{ImageMetadataList} object}
}
\value{
The number of image series.
}
\description{
Get the number of image series
}
\examples{
f = mockFile(sizeX = 256, sizeY = 256, series = 2)
img = read.image(f)
seriesCount(img)

meta <- metadata(img)
seriesCount(meta)
}
\author{
Andrzej Oleś <\email{andrzej.oles@gmail.com}>, 2014-2022
}
