% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedImage.R
\docType{class}
\name{AnnotatedImageList-class}
\alias{AnnotatedImageList-class}
\alias{AnnotatedImageList}
\title{AnnotatedImageList Class}
\usage{
AnnotatedImageList(...)
}
\arguments{
\item{...}{a list of \linkS4class{AnnotatedImage} objects to include in the new object.}
}
\description{
A list of \linkS4class{AnnotatedImage} objects.
}
