% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BFImage-class}
\alias{BFImage}
\alias{BFImage-class}
\alias{as.Image.BFImage}
\alias{coreMetadata}
\alias{globalMetadata}
\alias{seriesMetadata}
\alias{show,BFImage-method}
\title{Bio-Formats Image Class}
\usage{
BFImage(image = Image(), metadata = list(), omexml = "")

\S4method{show}{BFImage}(object)

coreMetadata(y)

globalMetadata(y)

seriesMetadata(y)

\method{as.Image}{BFImage}(x)
}
\arguments{
\item{image}{an Image object}

\item{metadata}{named list containing image metadata}

\item{omexml}{a string containing a dumped OME-XML DOM tree}

\item{object}{A BFImage object}

\item{y}{A BFImage object}

\item{x}{BFImage objects}
}
\value{
Named list consisting of key value pairs.
}
\description{
Extends the \code{\link[EBImage]{Image}} class from the \pkg{EBImage} package.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{named list containing image metadata}

\item{\code{omexml}}{a string containing a dumped OME-XML DOM tree}
}}
\examples{
img = read.image(mockFile())

coreMetadata(img)
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

