% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{BFImage-class}
\alias{BFImage}
\alias{BFImage-class}
\alias{coreMetadata}
\alias{show,BFImage-method}
\title{Bio-Formats Image Class}
\usage{
BFImage(image = Image(), coreMetadata = list(), globalMetadata = list(),
  seriesMetadata = list(), omexml = "")

\S4method{show}{BFImage}(object)

coreMetadata(y)
}
\arguments{
\item{image}{an Image object}

\item{coreMetadata}{named list containing image metadata}

\item{globalMetadata}{named list containing image metadata}

\item{seriesMetadata}{named list containing image metadata}

\item{omexml}{a string containing a dumped OME-XML DOM tree}

\item{y}{A BFImage object}

\item{object}{Any R object}
}
\value{
Named list consisting of key value pairs.
}
\description{
Extends the \code{\link[EBImage]{Image}} class from the \pkg{EBImage} package.
}
\section{Slots}{

\describe{
\item{\code{coreMetadata}}{named list containing image metadata}

\item{\code{globalMetadata}}{named list containing image metadata}

\item{\code{seriesMetadata}}{named list containing image metadata}

\item{\code{omexml}}{a string containing a dumped OME-XML DOM tree}
}}
\examples{
img = read.image(mockFile())

coreMetadata(img)
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

