% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BFImage.R
\docType{class}
\name{BFImage-class}
\alias{BFImage}
\alias{BFImage-class}
\alias{as.Image.BFImage}
\alias{coreMetadata}
\alias{globalMetadata}
\alias{print.BFImage}
\alias{seriesMetadata}
\alias{show,BFImage-method}
\title{Bio-Formats Image Class}
\usage{
BFImage(image = Image(), metadata = ImageMetadata(), omexml = "")

\S4method{show}{BFImage}(object)

\method{print}{BFImage}(x, short = FALSE, ...)

coreMetadata(y, series = 1L, ...)

globalMetadata(y, series = 1L, ...)

seriesMetadata(y, series = 1L)

seriesMetadata(y, series = 1L)

\method{as.Image}{BFImage}(x)
}
\arguments{
\item{image}{an Image object}

\item{metadata}{an ImageMetadata object containing image metadata}

\item{omexml}{a string containing a dumped OME-XML DOM tree}

\item{object}{A BFImage object}

\item{x}{a BFImage object}

\item{short}{turns off image data preview}

\item{...}{further arguments to be passed to other methods}

\item{y}{A BFImage object or metadata list}

\item{series}{series ID}

\item{arguments}{passed to \code{grep}}
}
\value{
Named list consisting of key value pairs.
}
\description{
Extends the \code{\link[EBImage]{Image}} class from the \pkg{EBImage} package.
}
\section{Slots}{

\describe{
\item{\code{omexml}}{a string containing a dumped OME-XML DOM tree}
}}
\examples{
img = read.image(mockFile())

coreMetadata(img)
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

