% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dimorder}
\alias{dimorder}
\title{Get the ordering of the image frames}
\usage{
dimorder(x)
}
\arguments{
\item{x}{An Image object or an array}
}
\value{
A character vector giving the dimension labels.
}
\description{
Get the ordering of the image frames
}

