% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotatedImage.R
\name{dimorder}
\alias{dimorder}
\title{Image Frames Order}
\usage{
dimorder(x)
}
\arguments{
\item{x}{An \code{\link[EBImage]{Image}} object or an array}
}
\value{
A character vector giving the dimension names.
}
\description{
Get the ordering of image frames.
}
\examples{
# sample timelapse image
f = mockFile(sizeC=2, sizeT=10)
img = read.image(f)

dimorder(img)
}
\author{
Andrzej Oles <\email{andrzej.oles@embl.de}>, 2014-2015
}

