% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImageMetadata.R
\docType{class}
\name{ImageMetadata-class}
\alias{ImageMetadata}
\alias{ImageMetadata-class}
\alias{print.ImageMetadata}
\alias{show,ImageMetadata-method}
\title{ImageMetadata Class}
\usage{
ImageMetadata(metadata = list())

\S4method{show}{ImageMetadata}(object)

\method{print}{ImageMetadata}(x, ...)
}
\arguments{
\item{metadata}{list containing image metadata}

\item{object}{An ImageMetadata object}

\item{object}{An ImageMetadata object}
}
\description{
Formal representation of image metadata.
}

