% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageMetadata.R
\docType{class}
\name{ImageMetadata-class}
\alias{ImageMetadata}
\alias{ImageMetadata-class}
\alias{ImageMetadataList}
\alias{ImageMetadataList-class}
\alias{print.ImageMetadata}
\alias{print.ImageMetadataList}
\alias{show,ImageMetadata-method}
\alias{show,ImageMetadataList-method}
\title{ImageMetadata and ImageMetadataList Class}
\usage{
ImageMetadata(...)

ImageMetadataList(...)

\method{print}{ImageMetadata}(x, list.len = 5L, ...)

\S4method{show}{ImageMetadata}(object)

\S4method{show}{ImageMetadataList}(object)

\method{print}{ImageMetadataList}(x, ...)
}
\arguments{
\item{...}{further arguments to be passed to other methods}

\item{x}{An ImageMetadata object}

\item{list.len}{numeric; maximum number of metadata entries to display}

\item{object}{an \code{ImageMetadata} object}
}
\description{
Formal representation of image metadata.
}
\author{
Andrzej Oleś <\email{andrzej.oles@embl.de}>, 2014-2017
}

