% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata.BFImage}
\alias{metadata.default}
\title{Image Metadata}
\usage{
metadata(x)

\method{metadata}{default}(x)

\method{metadata}{BFImage}(x)
}
\arguments{
\item{x}{Object}
}
\description{
Image Metadata
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

