% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataAccessors.R
\name{coreMetadata}
\alias{coreMetadata}
\alias{globalMetadata}
\alias{seriesMetadata}
\title{Metadata Accessors}
\usage{
coreMetadata(x, series, ...)

globalMetadata(x, series, ...)

seriesMetadata(x, series, ...)
}
\arguments{
\item{x}{an \linkS4class{AnnotatedImage}, \linkS4class{AnnotatedImageList}, \linkS4class{ImageMetadata}, or \linkS4class{ImageMetadataList} object}

\item{series}{series ID}

\item{...}{arguments passed to \code{\link[base]{grep}}}
}
\value{
Named list consisting of key value pairs.
}
\description{
Get image metadata.
}
\details{
The \ldots arguments are passed to \code{\link[base]{grep}} called on metadata names allowing for convenient subsetting.
}
\examples{
img = read.image(system.file("images", "nuclei.tif", package="EBImage"))
 
coreMetadata(img)

# subset for specific names 

globalMetadata(img, pattern="Image")
}
\author{
Andrzej Oleś <\email{andrzej.oles@embl.de}>, 2014-2015
}

