% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBioFormats.R, R/utils.R
\docType{package}
\name{RBioFormats}
\alias{RBioFormats}
\alias{BioFormats.version}
\title{R Interface to Bio-Formats}
\usage{
BioFormats.version()
}
\value{
A character string containing the Bio-Formats library version.
}
\description{
Interfaces the \href{http://www.openmicroscopy.org/bio-formats}{Bio-Formats} Java library.
}
\section{Functions}{
\itemize{
\item \code{BioFormats.version()}: Provides the version of the Bio-Formats library embedded in the package.

}}
\examples{
BioFormats.version()
}
\references{
\href{http://www.openmicroscopy.org/bio-formats}{Bio-Formats website}
}
\author{
Andrzej Oleś <\email{andrzej.oles@gmail.com}>, 2014-2022
}
