% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RBioFormats.R
\docType{package}
\name{RBioFormats}
\alias{RBioFormats}
\alias{RBioFormats-package}
\title{R Interface to Bio-Formats}
\description{
Interfaces the \href{http://www.openmicroscopy.org/site/products/bio-formats}{Bio-Formats} Java library.
}
\author{
Andrzej Oles <\email{andrzej.oles@embl.de}>, 2014-2015
}
\references{
\href{http://www.openmicroscopy.org/site/products/bio-formats}{Bio-Formats website}
}

