% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.image}
\alias{read.image}
\title{Read Images}
\usage{
read.image(file, filter.metadata = FALSE, proprietary.metadata = TRUE,
  normalize = TRUE)
}
\arguments{
\item{file}{character, file name}

\item{filter.metadata}{logical, specifies whether ugly metadata (entries with unprintable characters, and extremely large entries) should be discarded from the metadata table}

\item{proprietary.metadata}{logical, should proprietary metadata be populated to OME-XML}

\item{normalize}{logical, should the original image data be mapped to the [0,1] range}
}
\value{
A \code{\link{BFImage-class}} object.
}
\description{
Read image files using the Bio-Formats library. A list of supported formats can be found on the \href{http://www.openmicroscopy.org/site/support/bio-formats5/supported-formats.html}{Bio-Formats website}.
}
\examples{
require(EBImage)
f = system.file("images", "lena-color.png", package="EBImage")

img = read.image(f)
img
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

