% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.metadata}
\alias{read.metadata}
\title{Read Image Metadata}
\usage{
read.metadata(file, filter.metadata = FALSE)
}
\arguments{
\item{file}{character, file name}

\item{filter.metadata}{logical, specifies whether ugly metadata (entries with unprintable characters, and extremely large entries) should be discarded from the metadata table}
}
\value{
A list containing image meta-data.
}
\description{
Read image metadata using the Bio-Formats library. A list of supported formats can be found on the \href{http://www.openmicroscopy.org/site/support/bio-formats5/supported-formats.html}{Bio-Formats website}.
}
\examples{
require(EBImage)
f = system.file("images", "nuclei.tif", package="EBImage")

metadata = read.metadata(f)
str(metadata)
}
\author{
Andrzej Oles \email{andrzej.oles@embl.de}, 2014
}

