% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metadataAccessors.R
\name{seriesCount}
\alias{seriesCount}
\title{Number of Image Series}
\usage{
seriesCount(x)
}
\arguments{
\item{x}{an \code{AnnotatedImage} or \code{ImageMetadataList} object}
}
\description{
Get the number of image series
}
\author{
Andrzej Oles <\email{andrzej.oles@embl.de}>, 2014-2015
}

